+++ test/test_tool.py
@@ -32,9 +32,11 @@ class MockRMFNode:
     def get_name(self): return self.name
     def get_index(self): return self.index
 
-def make_node(name, index):
+def make_node(name, index, resolution=None):
     n = MockRMFNode(name, index)
-    return src.io._RMFHierarchyNode(n)
+    h = src.io._RMFHierarchyNode(n)
+    h.resolution = resolution
+    return h
 
 def make_feature(name, index):
     n = MockRMFNode(name, index)
@@ -48,7 +50,9 @@ def make_provenance(name, index):
 class Tests(unittest.TestCase):
     def test_rmf_hierarchy_model_none(self):
         """Test RMFHierarchyModel class with null hierarchy"""
-        m = src.tool._RMFHierarchyModel(None)
+        resolutions = set((1.0, 10.0))
+
+        m = src.tool._RMFHierarchyModel(None, resolutions)
         self.assertEqual(m.columnCount(None), 1)
 
         self.assertEqual(m.rowCount(QModelIndex()), 0)
@@ -64,8 +68,9 @@ class Tests(unittest.TestCase):
         grandchild = make_node("grandchild", 3)
         child2.add_children([grandchild])
         root.add_children((child1, child2))
+        resolutions = set((None,))
 
-        m = src.tool._RMFHierarchyModel(root)
+        m = src.tool._RMFHierarchyModel(root, resolutions)
         self.assertEqual(m.columnCount(None), 1)
         # Top level has one child (RMF root)
         self.assertEqual(m.rowCount(QModelIndex()), 1)
@@ -167,6 +172,8 @@ class Tests(unittest.TestCase):
         m1.rmf_hierarchy = None
         m1.rmf_features = []
         m1.rmf_provenance = []
+        m1._rmf_resolutions = set((1.0, 10.0))
+        m1._selected_rmf_resolutions = set((1.0, 10.0, None))
         m2 = Model(mock_session, 'test')
         mock_session.models.add((m1, m2))
         r = src.tool.RMFViewer(mock_session, "RMF Viewer")
@@ -175,6 +182,8 @@ class Tests(unittest.TestCase):
         m3.rmf_hierarchy = None
         m3.rmf_features = []
         m3.rmf_provenance = []
+        m3._rmf_resolutions = set((1.0, 10.0))
+        m3._selected_rmf_resolutions = set((1.0, 10.0, None))
         mock_session.models.add((m3,))
 
     @unittest.skipIf(utils.no_gui, "Cannot test without GUI")
@@ -206,9 +215,11 @@ class Tests(unittest.TestCase):
             for w in stack.widget(1).children():
                 if isinstance(w, QPushButton):
                     yield w
-        root = make_node("root", 0)
+        class TestChimeraObj:
+            pass
+        root = make_node("root", 0, resolution=10)
         child1 = make_node("child1", 1)
-        child1.chimera_obj = 'test object'
+        child1.chimera_obj = TestChimeraObj()
         child2 = make_node("child2", 2)
         grandchild = make_node("grandchild", 3)
         child2.add_children([grandchild])
@@ -220,16 +231,19 @@ class Tests(unittest.TestCase):
 
         m1.rmf_features = [make_node("f1", 4), make_node("f2", 5)]
         m1.rmf_provenance = []
+        m1._rmf_resolutions = set((1, 10))
+        m1._selected_rmf_resolutions = set((1, 10, None))
         mock_session.models.add((m1,))
         r = src.tool.RMFViewer(mock_session, "RMF Viewer")
         tree1 = get_first_tree(r.model_stack.widget(0))
         buttons = list(get_buttons(r.model_stack.widget(0)))
         # Show, View, Hide, Select
-        self.assertEqual(len(buttons), 4)
+        self.assertEqual(len(buttons), 5)
         # Call "clicked" methods directly
         r._select_button_clicked(tree1)
         tree1.selectAll()
         r._show_button_clicked(tree1)
+        r._show_only_button_clicked(tree1)
         r._hide_button_clicked(tree1)
         r._view_button_clicked(tree1)
         # Call indirectly via clicking each button
@@ -253,6 +267,8 @@ class Tests(unittest.TestCase):
         m1.rmf_features = [make_node("f1", 4), make_node("f2", 5)]
         m1.rmf_features[0].chimera_obj = 'test object'
         m1.rmf_provenance = []
+        m1._rmf_resolutions = set((1.0, 10.0))
+        m1._selected_rmf_resolutions = set((1.0, 10.0, None))
         mock_session.models.add((m1,))
         r = src.tool.RMFViewer(mock_session, "RMF Viewer")
         tree1 = get_first_tree(r.model_stack.widget(0))
@@ -300,6 +316,8 @@ class Tests(unittest.TestCase):
         m1.rmf_hierarchy = None
         m1.rmf_features = []
         m1.rmf_provenance = []
+        m1._rmf_resolutions = set((1.0, 10.0))
+        m1._selected_rmf_resolutions = set((1.0, None))
         mock_session.models.add((m1,))
         r = src.tool.RMFViewer(mock_session, "RMF Viewer")
         s = r.take_snapshot(mock_session, None)